recsys_api/requirements.txt
orjson==3.9.15
kafka-python==2.0.2
backoff==2.2.1
pydantic==2.6.3
pydantic-settings==2.2.1
